add validation user delete

+++ app/Actions/UserAction.php
@@ -6,6 +6,7 @@ use App\Models\Solver;
 use App\Models\User;
 use App\Models\TeamManage;
 use App\Models\UserRole;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Hash;
 
@@ -16,17 +17,20 @@ class UserAction
         return User::get();
     }
 
-    public function getByUsername($username){
+    public function getByUsername($username)
+    {
         $user = User::with('solvers.challenge.category')->where('users.username', $username)->get();
-        return $user; 
+        return $user;
     }
 
-    public function getUserById($id){
+    public function getUserById($id)
+    {
         $user = User::with('role.role')->find($id);
         return $user;
     }
 
-    public function storeUser($request, UserRoleAction $roleAction){
+    public function storeUser($request, UserRoleAction $roleAction)
+    {
         $id = Str::uuid()->toString();
         $user = new User();
 
@@ -39,17 +43,21 @@ class UserAction
         $roleAction->storeUserRole($id);
     }
 
-    public function updateUser($request, $id){
+    public function updateUser($request, $id)
+    {
         $user = User::find($id);
         $user->username = $request['username'];
         $user->name = $request['name'];
         $user->save();
     }
 
-    public function deleteUser($id){
-        Solver::where('solvers.user_id', $id)->delete();
-        TeamManage::where('user_id', $id)->delete();
-        UserRole::where('user_id', $id)->delete();
-        User::find($id)->delete();
+    public function deleteUser($id)
+    {
+        if (!(Auth::user()->id === $id)) {
+            Solver::where('solvers.user_id', $id)->delete();
+            TeamManage::where('user_id', $id)->delete();
+            UserRole::where('user_id', $id)->delete();
+            User::find($id)->delete();
+        }
     }
 }
